backend/functions/requirements.txt
azure-functions==1.17.0
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
pyodbc==4.0.39
requests==2.31.0
python-dotenv==1.0.0